* commit 'a21039c3ae2b20e44ceb2735251c04d0aac89afd':
  Fix a bunch more lint.

+++ src/common_test.h
@@ -49,11 +49,11 @@ static const byte kBase64Map[256] = {
   255, 255, 255, 255, 255, 255, 255,  62, 255, 255, 255,  63,
   52,  53,  54,  55,  56,  57,  58,  59,  60,  61, 255, 255,
   255, 254, 255, 255, 255,   0,   1,   2,   3,   4,   5,   6,
-    7,   8,   9,  10,  11,  12,  13,  14,  15,  16,  17,  18,
-   19,  20,  21,  22,  23,  24,  25, 255, 255, 255, 255, 255,
+    7,   8,   9,  10,  11,  12,  13,  14,  15,  16,  17,  18, // NOLINT
+   19,  20,  21,  22,  23,  24,  25, 255, 255, 255, 255, 255, // NOLINT
   255,  26,  27,  28,  29,  30,  31,  32,  33,  34,  35,  36,
-   37,  38,  39,  40,  41,  42,  43,  44,  45,  46,  47,  48,
-   49,  50,  51, 255, 255, 255, 255, 255, 255, 255, 255, 255,
+   37,  38,  39,  40,  41,  42,  43,  44,  45,  46,  47,  48, // NOLINT
+   49,  50,  51, 255, 255, 255, 255, 255, 255, 255, 255, 255, // NOLINT
   255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
   255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
   255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
@@ -69,7 +69,7 @@ static const byte kBase64Map[256] = {
 
 byte* DecodeBase64(const char* src, size_t* dst_size) {
   std::vector<byte> tmp;
-  unsigned long t = 0, y = 0;
+  uint32_t t = 0, y = 0;
   int g = 3;
   for (size_t i = 0; src[i] != '\0'; ++i) {
     byte c = kBase64Map[src[i] & 0xFF];
@@ -172,7 +172,6 @@ class ScratchFile {
 
 class CommonTest : public testing::Test {
  public:
-
   static void MakeExecutable(const ByteArray* code_array) {
     CHECK(code_array != NULL);
     MakeExecutable(code_array->GetData(), code_array->GetLength());
